Mixin global context asset when rendering on server
Add the concept of a global context asset. This is a special asset that
contains extensions to `HAML.globals`. This asset is mixed into the
javascript environment before compilation on the server - allowing us
to share custom `HAML.globals` on both client and server.

The path for this asset is configurable.