Monkey patch rails: bust server template cache in development
When the global_context_asset changes rails should recompile stale hamlc
templates. By default, rails only recompiles templates if they're
touched, which is problematic since it's the helper not the template that
changed.

This commit patches rails so that it automatically recompiles and touches
hamlc templates when the global_context_asset changes from underneath it.